@@ -16,12 +16,12 @@
 		}
 	],
 	"require": {
-		"php": ">= 7.3",
+		"php": ">= 8.0",
 		"ext-json": "*",
 		"nette/application": "^3.1",
 		"nette/di": "^3.0",
 		"nette/utils": "^3.2|^4.0",
-		"latte/latte": "^2.9|^3.0"
+		"latte/latte": "^3.0"
 	},
 	"archive": {
 		"exclude": [
